fixing merge

+++ lib/data/models/response/auth_response/dashboard_content_resp.dart
@@ -1327,7 +1327,7 @@ class DashboardMyCoursesLimit {
   String? name;
   int? hours;
   int? enrollments;
-  int? completion;
+  double? completion;
 
   DashboardMyCoursesLimit({
     this.id,
@@ -1344,7 +1344,7 @@ class DashboardMyCoursesLimit {
     name = json['name'] as String?;
     hours = json['hours'] as int?;
     enrollments = json['enrollments'] as int?;
-    completion = json['completion'] ;
+    completion = json['completion'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
